Switch to integer clipping distances (#3666)
This removes the last floating point math from clipping. Roughly, this:
* Switches to use L_inf distances instead of L_2.
* Switches length_scale to not try to be right at percolation
criticality. In practice, I think this means that our cutoff distance
will end up being like 2.5x criticality instead of 3x criticality.
* Uses an integer square root.

I've done a smoke test of confirming that clipping still happens, but
not that the clipping feels great. I think the latter can be addressed
by adjusting scaled_cutoff_distance if needed. But I think "integer
math" is the right place from which to iterate.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211917043032500)

+++ packages/mettagrid/python/src/mettagrid/mettagrid_c.pyi
@@ -117,7 +117,7 @@ class ClipperConfig:
     def __init__(
         self) -> None: ...
     unclipping_protocols: list[Protocol]
-    length_scale: float
+    length_scale: int
     scaled_cutoff_distance: int
     clip_period: int
